File_Service/requirements.txt
minio
redis
grpcio
protobuf